refactor(package): menuGroup 관련 로직들 menuGroup 패키지로 나눔 및 menu 내 menuGroup 결합도 낮춤
menu
- menu에서 menuGroup 넣는 이유는, menuGroup이 비었는지 validation 하기 때문이므로, 이를 따로 validation 로직으로 빼고 id로 저장함

menuGroup 관련 로직
- 패키지 리팩토링
