+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -3,6 +3,7 @@ package kitchenpos.menu.domain;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
+import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -15,6 +16,7 @@ import javax.persistence.ManyToOne;
 import kitchenpos.common.constant.ErrorCode;
 import kitchenpos.common.domain.Name;
 import kitchenpos.common.domain.Price;
+import kitchenpos.menugroup.domain.MenuGroup;
 
 @Entity
 public class Menu {
@@ -26,41 +28,33 @@ public class Menu {
     private Name name;
     @Embedded
     private Price price;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "menu_group_id", foreignKey = @ForeignKey(name = "fk_menu_to_menu_group"))
-    private MenuGroup menuGroup;
+    @Column(nullable = false)
+    private Long menuGroupId;
     @Embedded
     private MenuProducts menuProducts;
 
     protected Menu() {}
 
-    private Menu(Long id, String name, BigDecimal bigDecimalPrice, MenuGroup menuGroup,
+    private Menu(Long id, String name, BigDecimal bigDecimalPrice, Long menuGroupId,
                  MenuProducts menuProducts) {
         Price price = Price.from(bigDecimalPrice);
-        validateMenuGroup(menuGroup);
         validateMenuProducts(price, menuProducts);
         menuProducts.setUpMenu(this);
         this.id = id;
         this.name = Name.from(name);
         this.price = price;
-        this.menuGroup = menuGroup;
+        this.menuGroupId = menuGroupId;
         this.menuProducts = menuProducts;
     }
 
-    public static Menu of(Long id, String name, BigDecimal price, MenuGroup menuGroup,
+    public static Menu of(Long id, String name, BigDecimal price, Long menuGroupId,
                           List<MenuProduct> menuProducts) {
-        return new Menu(id, name, price, menuGroup, MenuProducts.from(menuProducts));
+        return new Menu(id, name, price, menuGroupId, MenuProducts.from(menuProducts));
     }
 
-    public static Menu of(String name, BigDecimal price, MenuGroup menuGroup,
+    public static Menu of(String name, BigDecimal price, Long menuGroupId,
                           MenuProducts menuProducts) {
-        return new Menu(null, name, price, menuGroup, menuProducts);
-    }
-
-    private void validateMenuGroup(MenuGroup menuGroup) {
-        if(menuGroup == null) {
-            throw new IllegalArgumentException(ErrorCode.메뉴_그룹은_비어있을_수_없음.getErrorMessage());
-        }
+        return new Menu(null, name, price, menuGroupId, menuProducts);
     }
 
     private void validateMenuProducts(Price price, MenuProducts menuProducts) {
@@ -81,8 +75,8 @@ public class Menu {
         return price;
     }
 
-    public MenuGroup getMenuGroup() {
-        return menuGroup;
+    public Long getMenuGroupId() {
+        return menuGroupId;
     }
 
     public MenuProducts getMenuProducts() {
