+++ src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProducts;
 
 public class MenuRequest {
@@ -24,7 +24,7 @@ public class MenuRequest {
     }
 
     public Menu toMenu(MenuGroup menuGroup, MenuProducts menuProducts) {
-        return Menu.of(name, price, menuGroup, menuProducts);
+        return Menu.of(name, price, menuGroup.getId(), menuProducts);
     }
 
     public String getName() {
