+++ src/main/java/kitchenpos/menu/dto/MenuResponse.java
@@ -4,7 +4,6 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.common.domain.Name;
 import kitchenpos.common.domain.Price;
 
@@ -18,11 +17,11 @@ public class MenuResponse {
 
     private MenuResponse() {}
 
-    private MenuResponse(Long id, Name name, Price price, MenuGroup menuGroup, List<MenuProductResponse> menuProductResponses) {
+    private MenuResponse(Long id, Name name, Price price, Long menuGroupId, List<MenuProductResponse> menuProductResponses) {
         this.id = id;
         this.name = name.value();
         this.price = price.value();
-        this.menuGroupId = menuGroup.getId();
+        this.menuGroupId = menuGroupId;
         this.menuProductResponses = menuProductResponses;
     }
 
@@ -32,7 +31,7 @@ public class MenuResponse {
                 .stream()
                 .map(MenuProductResponse::from)
                 .collect(Collectors.toList());
-        return new MenuResponse(menu.getId(), menu.getName(), menu.getPrice(), menu.getMenuGroup(), menuProductResponses);
+        return new MenuResponse(menu.getId(), menu.getName(), menu.getPrice(), menu.getMenuGroupId(), menuProductResponses);
     }
 
     public Long getId() {
