+++ src/main/java/kitchenpos/menugroup/dto/MenuGroupResponse.java
@@ -1,6 +1,6 @@
-package kitchenpos.menu.dto;
+package kitchenpos.menugroup.dto;
 
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.common.domain.Name;
 
 public class MenuGroupResponse {
