+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -17,10 +17,9 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import kitchenpos.common.constant.ErrorCode;
-import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuProducts;
 import kitchenpos.menu.domain.MenuRepository;
