+++ src/test/java/kitchenpos/menu/domain/MenuGroupTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import kitchenpos.common.constant.ErrorCode;
 import kitchenpos.common.domain.Name;
+import kitchenpos.menugroup.domain.MenuGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
