+++ src/test/java/kitchenpos/menu/domain/MenuTestFixture.java
@@ -2,22 +2,23 @@ package kitchenpos.menu.domain;
 
 import java.math.BigDecimal;
 import java.util.List;
+import kitchenpos.common.constant.ErrorCode;
 import kitchenpos.common.domain.Name;
 import kitchenpos.common.domain.Price;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menugroup.domain.MenuGroup;
 
 public class MenuTestFixture {
 
     public static Menu generateMenu(Long id, String name, BigDecimal price, MenuGroup menuGroup, List<MenuProduct> menuProducts) {
-        return Menu.of(id, name, price, menuGroup, menuProducts);
+        validateMenuGroup(menuGroup);
+        return Menu.of(id, name, price, menuGroup.getId(), menuProducts);
     }
 
     public static Menu generateMenu(String name, BigDecimal price, MenuGroup menuGroup, List<MenuProduct> menuProducts) {
-        return Menu.of(name, price, menuGroup, MenuProducts.from(menuProducts));
+        validateMenuGroup(menuGroup);
+        return Menu.of(name, price, menuGroup.getId(), MenuProducts.from(menuProducts));
     }
 
     public static MenuRequest generateMenuRequest(Name name, Price price, Long menuGroupId, List<MenuProductRequest> menuProductRequests) {
@@ -30,5 +31,11 @@ public class MenuTestFixture {
 
     public static MenuRequest generateMenuRequest(String name, BigDecimal price, Long menuGroupId, List<MenuProductRequest> menuProductRequests) {
         return new MenuRequest(name, price, menuGroupId, menuProductRequests);
+    }
+
+    private static void validateMenuGroup(MenuGroup menuGroup) {
+        if(menuGroup == null) {
+            throw new IllegalArgumentException(ErrorCode.메뉴_그룹은_비어있을_수_없음.getErrorMessage());
+        }
     }
 }
