+++ src/test/java/kitchenpos/order/domain/OrderTableTest.java
@@ -1,6 +1,5 @@
 package kitchenpos.order.domain;
 
-import static java.util.Collections.singletonList;
 import static kitchenpos.menu.domain.MenuGroupTestFixture.generateMenuGroup;
 import static kitchenpos.menu.domain.MenuProductTestFixture.generateMenuProduct;
 import static kitchenpos.menu.domain.MenuTestFixture.generateMenu;
@@ -12,8 +11,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.product.domain.Product;
