+++ src/test/java/kitchenpos/order/domain/OrderTest.java
@@ -16,7 +16,7 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import kitchenpos.common.constant.ErrorCode;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.product.domain.Product;
